requirements.txt
altair
numpy
pandas
pydeck
streamlit
google-api-python-client